# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MusicDrop",
    "version": "dev"
  },
  "paths": {
...
    "/api/albums": {
      "get": {
        "tags": [
          "albums"
        ],
        "summary": "List Albums Endpoint",
        "operationId": "list_albums_endpoint_api_albums_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "artist",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Artist"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlbumPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/albums/{album_id}": {
      "get": {
        "tags": [
          "albums"
        ],
        "summary": "Get Album Detail Endpoint",
        "operationId": "get_album_detail_endpoint_api_albums__album_id__get",
        "parameters": [
          {
            "name": "album_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Album Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlbumDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "albums"
        ],
        "summary": "Delete Album Endpoint",
        "description": "Move the album's whole folder to Trash (reversible) and drop it from the\nlibrary. 404 unknown album; 409 while a library job is running.",
        "operationId": "delete_album_endpoint_api_albums__album_id__delete",
        "parameters": [
          {
            "name": "album_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Album Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/albums/{album_id}/missing": {
      "get": {
        "tags": [
          "albums"
        ],
        "summary": "Get Album Missing Endpoint",
        "description": "Full release tracklist vs. the library (missing rows + counts). Read-only.",
        "operationId": "get_album_missing_endpoint_api_albums__album_id__missing_get",
        "parameters": [
          {
            "name": "album_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Album Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlbumMissingReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/albums/{album_id}/edit/preview": {
      "post": {
        "tags": [
          "albums"
        ],
        "summary": "Preview Album Edit Endpoint",
        "description": "Preview a pending album/track edit (field diff + move plan). Read-only.",
        "operationId": "preview_album_edit_endpoint_api_albums__album_id__edit_preview_post",
        "parameters": [
          {
            "name": "album_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Album Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlbumEditRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlbumEditPreview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/albums/{album_id}/edit": {
      "post": {
        "tags": [
          "albums"
        ],
        "summary": "Edit Album Endpoint",
        "description": "Apply an album/track tag edit (write + config-gated move). 409 if importing.",
        "operationId": "edit_album_endpoint_api_albums__album_id__edit_post",
        "parameters": [
          {
            "name": "album_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Album Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlbumEditRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlbumEditResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/albums/{album_id}/cover": {
      "get": {
        "tags": [
          "albums"
        ],
        "summary": "Get Album Cover Endpoint",
        "operationId": "get_album_cover_endpoint_api_albums__album_id__cover_get",
        "parameters": [
          {
            "name": "album_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Album Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "albums"
        ],
        "summary": "Install Album Cover Endpoint",
        "description": "Install an uploaded (or approved-fetched) cover. 409 while importing.",
        "operationId": "install_album_cover_endpoint_api_albums__album_id__cover_post",
        "parameters": [
          {
            "name": "album_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Album Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_install_album_cover_endpoint_api_albums__album_id__cover_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoverInstallResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/albums/{album_id}/cover/fetch": {
      "post": {
        "tags": [
          "albums"
        ],
        "summary": "Fetch Album Cover Endpoint",
        "description": "Fetch beets' best cover candidate. Returns the image (preview) or 404. No write.",
        "operationId": "fetch_album_cover_endpoint_api_albums__album_id__cover_fetch_post",
        "parameters": [
          {
            "name": "album_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Album Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/albums/{album_id}/lyrics/fetch": {
      "post": {
        "tags": [
          "albums"
        ],
        "summary": "Fetch Album Lyrics Endpoint",
        "description": "Start an album-scoped lyrics fetch job (writes tags \u2192 Plex). Poll\nGET /api/lyrics/backfill for marching progress. 404 unknown album, 409 if busy.",
        "operationId": "fetch_album_lyrics_endpoint_api_albums__album_id__lyrics_fetch_post",
        "parameters": [
          {
            "name": "album_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Album Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LyricsBackfillStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Album": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "album_artist": {
            "type": "string",
            "title": "Album Artist"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          },
          "track_count": {
            "type": "integer",
            "title": "Track Count"
          },
          "genre": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Genre"
          },
          "mb_albumid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mb Albumid"
          }
        },
        "type": "object",
        "required": [
          "id",
          "album_artist",
          "title",
          "year",
          "track_count",
          "genre",
          "mb_albumid"
        ],
        "title": "Album"
      },
...
      "AlbumDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "album_artist": {
            "type": "string",
            "title": "Album Artist"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          },
          "track_count": {
            "type": "integer",
            "title": "Track Count"
          },
          "genre": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Genre"
          },
          "mb_albumid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mb Albumid"
          },
          "tracks": {
            "items": {
              "$ref": "#/components/schemas/Track"
            },
            "type": "array",
            "title": "Tracks"
          },
          "release": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReleaseIdentity"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "album_artist",
          "title",
          "year",
          "track_count",
          "genre",
          "mb_albumid",
          "tracks"
        ],
        "title": "AlbumDetail"
      },
...
      "AlbumEditPreview": {
        "properties": {
          "changed_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Changed Fields"
          },
          "album_before": {
            "$ref": "#/components/schemas/AlbumDiffSide"
          },
          "album_after": {
            "$ref": "#/components/schemas/AlbumDiffSide"
          },
          "tracks": {
            "items": {
              "$ref": "#/components/schemas/EditTrackChange"
            },
            "type": "array",
            "title": "Tracks"
          },
          "move_enabled": {
            "type": "boolean",
            "title": "Move Enabled"
          },
          "move_plan": {
            "items": {
              "$ref": "#/components/schemas/TrackPathChange"
            },
            "type": "array",
            "title": "Move Plan"
          }
        },
        "type": "object",
        "required": [
          "changed_fields",
          "album_before",
          "album_after",
          "tracks",
          "move_enabled",
          "move_plan"
        ],
        "title": "AlbumEditPreview",
        "description": "The preview of a pending edit: field diff + move plan. Persists nothing."
      },
      "AlbumEditRequest": {
        "properties": {
          "album": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AlbumFieldEdits"
              },
              {
                "type": "null"
              }
            ]
          },
          "tracks": {
            "items": {
              "$ref": "#/components/schemas/TrackFieldEdits"
            },
            "type": "array",
            "title": "Tracks",
            "default": []
          }
        },
        "type": "object",
        "title": "AlbumEditRequest",
        "description": "One edit request: optional album-header changes + per-track changes."
      },
      "AlbumEditResult": {
        "properties": {
          "album": {
            "$ref": "#/components/schemas/AlbumDetail"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ItemWriteResult"
            },
            "type": "array",
            "title": "Items"
          },
          "write_failures": {
            "type": "integer",
            "title": "Write Failures"
          },
          "move_failures": {
            "type": "integer",
            "title": "Move Failures"
          }
        },
        "type": "object",
        "required": [
          "album",
          "items",
          "write_failures",
          "move_failures"
        ],
        "title": "AlbumEditResult",
        "description": "The result of an apply: the refreshed album + per-track outcomes."
      },
...
      "AlbumMissingReport": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ok",
              "no_musicbrainz_id",
              "release_unavailable",
              "fetch_failed"
            ],
            "title": "Status"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "present_count": {
            "type": "integer",
            "title": "Present Count"
          },
          "missing": {
            "items": {
              "$ref": "#/components/schemas/MissingReleaseTrack"
            },
            "type": "array",
            "title": "Missing"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          }
        },
        "type": "object",
        "required": [
          "status",
          "total",
          "present_count",
          "missing",
          "source"
        ],
        "title": "AlbumMissingReport"
      },
      "AlbumPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Album"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "AlbumPage"
      },
...
      "Body_install_album_cover_endpoint_api_albums__album_id__cover_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_install_album_cover_endpoint_api_albums__album_id__cover_post"
      },
...
      "CoverInstallResult": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "embedded": {
            "type": "boolean",
            "title": "Embedded"
          },
          "embed_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embed Detail"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "embedded"
        ],
        "title": "CoverInstallResult",
        "description": "Outcome of installing an album cover (set artpath + optional embed).\n\n``embedded`` is True only when the user's config enables ``embedart`` and\nthe image is jpeg/png; ``embed_detail`` explains why it was skipped."
      },
      "DeleteResult": {
        "properties": {
          "trashed_albums": {
            "type": "integer",
            "title": "Trashed Albums"
          },
          "trash_path": {
            "type": "string",
            "title": "Trash Path"
          }
        },
        "type": "object",
        "required": [
          "trashed_albums",
          "trash_path"
        ],
        "title": "DeleteResult",
        "description": "Outcome of a reversible delete: how many albums went to Trash + where.\n\n``trashed_albums`` is 1 for a single-album delete, N for an artist (every\nalbum of theirs). ``trash_path`` is the Trash location the files were moved\nto (recoverable from there)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LyricsBackfillStatus": {
        "properties": {
          "phase": {
            "type": "string",
            "enum": [
              "idle",
              "running",
              "done",
              "stopped",
              "failed"
            ],
            "title": "Phase"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "processed": {
            "type": "integer",
            "title": "Processed"
          },
          "found": {
            "type": "integer",
            "title": "Found"
          },
          "not_found": {
            "type": "integer",
            "title": "Not Found"
          },
          "failed": {
            "type": "integer",
            "title": "Failed"
          },
          "skipped": {
            "type": "integer",
            "title": "Skipped"
          },
          "current": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current"
          },
          "writes_enabled": {
            "type": "boolean",
            "title": "Writes Enabled"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "album_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Album Id"
          },
          "scope_label": {
            "type": "string",
            "title": "Scope Label"
          }
        },
        "type": "object",
        "required": [
          "phase",
          "job_id",
          "total",
          "processed",
          "found",
          "not_found",
          "failed",
          "skipped",
          "current",
          "writes_enabled",
          "error",
          "album_id",
          "scope_label"
        ],
        "title": "LyricsBackfillStatus"
      },
...
      "Track": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "track": {
            "type": "integer",
            "title": "Track"
          },
          "disc": {
            "type": "integer",
            "title": "Disc"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "artist": {
            "type": "string",
            "title": "Artist"
          },
          "mb_trackid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mb Trackid"
          },
          "has_lyrics": {
            "type": "boolean",
            "title": "Has Lyrics"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "track",
          "disc",
          "duration_seconds",
          "artist",
          "mb_trackid",
          "has_lyrics"
        ],
        "title": "Track"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}